Changed currentTime back to uint32_t
uint64_t was just causing problems and not doing any good

+++ Clock.h
@@ -62,9 +62,9 @@ public:
 		  uint8_t lightSensor, bool invertAmbientLightLevel, uint32_t minBrightness, uint32_t maxBrightness,
 		  uint8_t button, BrightnessChangeCB);
 	
-	uint64_t currentTime() { return _currentTime; }
+	uint32_t currentTime() { return _currentTime; }
 	std::string strftime(const char* format, uint32_t time) { return _timeWeatherServer.strftime(format, time); }
-	std::string prettyDay(uint64_t time) { return _timeWeatherServer.prettyDay(time); }
+	std::string prettyDay(uint32_t time) { return _timeWeatherServer.prettyDay(time); }
 	const char* weatherConditions() { return _timeWeatherServer.conditions(); }
 	uint32_t currentTemp() { return _timeWeatherServer.currentTemp(); }
 	uint32_t highTemp() { return _timeWeatherServer.highTemp(); }
@@ -83,7 +83,7 @@ private:
 	mil::BrightnessManager _brightnessManager;
 	Ticker _secondTimer;
 	
-	uint64_t _currentTime = 0;
+	uint32_t _currentTime = 0;
 	bool _needsUpdate = true;
 	
 	struct tm  _settingTime;
